long long to int64_t

+++ OnnxAttrs.cpp
@@ -42,9 +42,9 @@ template<> std::vector<int> OnnxAttrs::get<std::vector<int>>(std::string key) co
   auto attr = this->at(key)->ints();
   return std::vector<int>(attr.begin(), attr.end());
 }
-template<> std::vector<long long> OnnxAttrs::get<std::vector<long long>>(std::string key) const {
+template<> std::vector<int64_t> OnnxAttrs::get<std::vector<int64_t>>(std::string key) const {
   auto attr = this->at(key)->ints();
-  return std::vector<long long>(attr.begin(), attr.end());
+  return std::vector<int64_t>(attr.begin(), attr.end());
 }
 template<> std::vector<float> OnnxAttrs::get<std::vector<float>>(std::string key) const {
   auto attr = this->at(key)->floats();
